+++ retry/retry.go
@@ -7,52 +7,20 @@ import (
 	"github.com/blend/go-sdk/ex"
 )
 
-// Options are the options for the retry.
-type Options struct {
-	MaxAttempts         uint
-	DelayProvider       DelayProvider
-	ShouldRetryProvider ShouldRetryProvider
-}
-
-// Option mutates retry options.
-type Option func(*Options)
-
-// OptMaxAttempts sets the max attempts.
-func OptMaxAttempts(maxAttempts uint) Option {
-	return func(ro *Options) { ro.MaxAttempts = maxAttempts }
-}
-
-// OptDelayProvider sets the retry delay provider.
-func OptDelayProvider(delayProvider DelayProvider) Option {
-	return func(ro *Options) { ro.DelayProvider = delayProvider }
-}
-
-// OptConstantDelay sets the retry delay provider.
-func OptConstantDelay(d time.Duration) Option {
-	return func(ro *Options) { ro.DelayProvider = ConstantDelay(d) }
-}
-
-// OptExponentialDelay sets the retry delay provider.
-func OptExponentialDelay(d time.Duration) Option {
-	return func(ro *Options) { ro.DelayProvider = ExponentialBackoff(d) }
-}
-
-// OptShouldRetryProvider sets the should retry provider.
-func OptShouldRetryProvider(provider ShouldRetryProvider) Option {
-	return func(ro *Options) { ro.ShouldRetryProvider = provider }
-}
-
-// Retry calls a function and
-func Retry(ctx context.Context, action Action, opts ...Option) (res interface{}, err error) {
-	options := Options{
-		MaxAttempts:         DefaultMaxAttempts,
-		DelayProvider:       ConstantDelay(DefaultRetryDelay),
-		ShouldRetryProvider: func(_ error) bool { return true },
-	}
+// Retry calls a function and retries on error or if a should retry provider
+// is set, based on the should retry result.
+func Retry(ctx context.Context, action Action, opts ...Option) (interface{}, error) {
+	var options Options
+	DefaultOptions(&options)
 	for _, opt := range opts {
 		opt(&options)
 	}
+	return options.Retry(ctx, action)
+}
 
+// Retry calls a function and retries on error or if a should retry provider
+// is set, based on the should retry result.
+func (options Options) Retry(ctx context.Context, action Action) (res interface{}, err error) {
 	var attempt uint
 	for attempt = 0; attempt < options.MaxAttempts; attempt++ {
 		func() {
